Split up move arrows into separate files
