+++ modules/SynthesisCore/EntityMovement/ArrowBase.cs
@@ -1,0 +1,7 @@
+﻿namespace SynthesisCore.EntityMovement
+{
+    public abstract class ArrowBase
+    {
+        public abstract void Update();
+    }
+}
